refactor: WinnerCount 삭제
